fix(licensing): fix missed fourth tier-mapping implementation for compliance_reports (SMI-3140)
packages/enterprise/src/license/types.ts is a fourth, independent
tier->feature encoding (TeamFeatureFlag/EnterpriseFeatureFlag unions,
TEAM_FEATURES/ENTERPRISE_FEATURES arrays, TIER_FEATURES map) missed by
the prior commit's three-implementation sweep. It's load-bearing:
LicenseKeyGenerator.createTeamLicense() mints Team JWTs directly from
these arrays, and license.ts's checkFeature() enforces against the
JWT's features array independent of the tier-rank gate in
toolFeatureMapping.ts -- so a Team license key would still have been
denied compliance_reports even after the prior commit, exactly the
split-path failure pattern documented at license.tier.ts:76-80 (SMI-1953).

Found by an independent Fable-model review pass. Also updates the
FeatureChecker/LicenseKeyGenerator/LicenseValidator test files whose
assertions still encoded the old enterprise-only behavior, splits the
three test files that were already over the 500-line gate before this
change and got pushed further over by the fix, and adds the CHANGELOG
[Unreleased] entries required by CI's Check 54 (mcp-server, cli,
enterprise).
