+++ packages/enterprise/src/license/types.ts
@@ -25,6 +25,8 @@ export type TeamFeatureFlag =
   | 'usage_analytics'
   | 'priority_support'
   | 'skill_security_audit'
+  // SMI-3140: expanded to Team + Enterprise (2026-07-14)
+  | 'compliance_reports'
 
 /**
  * Enterprise tier feature flags (includes all Team features)
@@ -34,7 +36,6 @@ export type EnterpriseFeatureFlag =
   | 'rbac'
   | 'audit_logging'
   | 'siem_export'
-  | 'compliance_reports'
   | 'private_registry'
   | 'custom_integrations'
   | 'advanced_analytics'
@@ -59,6 +60,8 @@ export const TEAM_FEATURES: readonly TeamFeatureFlag[] = [
   'usage_analytics',
   'priority_support',
   'skill_security_audit',
+  // SMI-3140: expanded to Team + Enterprise (2026-07-14)
+  'compliance_reports',
 ] as const
 
 export const ENTERPRISE_FEATURES: readonly EnterpriseFeatureFlag[] = [
@@ -66,7 +69,6 @@ export const ENTERPRISE_FEATURES: readonly EnterpriseFeatureFlag[] = [
   'rbac',
   'audit_logging',
   'siem_export',
-  'compliance_reports',
   'private_registry',
   'custom_integrations',
   'advanced_analytics',
